requirements.txt
# EnglishLeader Bot — Render.com requirements
# Using >= for bulletproof installs on any Python 3.11+
python-telegram-bot>=21.0,<23.0
aiosqlite>=0.20.0
aiohttp>=3.10.0
APScheduler>=3.10.0
deep-translator>=1.11.0
gTTS>=2.5.0
pydub>=0.25.1
matplotlib>=3.8.0
Pillow>=10.0.0
langdetect>=1.0.9
python-dotenv>=1.0.0
PyPDF2>=3.0.1
python-docx>=1.0.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
requests>=2.31.0
cachetools>=5.3.0
